move glob to utils
