+++ src/Reader.cc
@@ -8,7 +8,7 @@
   #include "podio/SIOReader.h"
 #endif
 
-#include "globUtils.h"
+#include "podio/utilities/Glob.h"
 
 #include "TFile.h"
 #include "TKey.h"
@@ -21,7 +21,7 @@ Reader::Reader(std::unique_ptr<T> reader) : m_self(std::make_unique<ReaderModel<
 }
 
 Reader makeReader(const std::string& filename) {
-  return makeReader(detail::expand_glob(filename));
+  return makeReader(utils::expand_glob(filename));
 }
 
 Reader makeReader(const std::vector<std::string>& filenames) {
